feat: update review pages for better integration
