+++ src/app/home/pages/ListarReview.tsx
@@ -1,7 +1,7 @@
-import React, { useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import styles from "/src/app/home/styles/ListarReview.module.css";
 import { useNavigate } from "react-router-dom";
-import StarRating from "../../../shared/components/StarRating";
+import StarRating from "../../../shared/components/StarRating/StarRating";
 
 interface Review {
   id: number;
